Move tests of "lib/config" packages to the appropriate directory
